마지막 commit

+++ shop/views.py
@@ -38,23 +38,25 @@ shop_detail = DetailView.as_view(model=Shop)
 @login_required  #-> 장식자를 통해서 로그인 시에만 shop_new만 실행되도록 함
 def shop_new(request):
     form_cls = ShopForm
-    form_cls2 = ItemForm
+    #form_cls2 = ItemForm
 
     if request.method == "POST":   # GET 빈폼 or POST 반드시 입력 받음
         form = form_cls(request.POST, request.FILES, prefix="shop")
-        form2 = form_cls2(request.POST, request.FILES, prefix="item")
-        if form.is_valid() and form2.is_valid(): # 밸리드 체크 시 오류가 없을 경우에만 True리턴
-            shop = form.save() # 폼에 입력된 값을 저장
-            form2.save()
+        #form2 = form_cls2(request.POST, request.FILES, prefix="item")
+        if form.is_valid(): #and form2.is_valid(): # 밸리드 체크 시 오류가 없을 경우에만 True리턴
+            shop = form.save(commit=False) # 폼에 입력된 값을 저장, commit=FALSE를 사용하면 아래에 지정한 필드에 대해서는 save 제외됨
+            shop.user = request.user   # user 필드를 자동으로 세팅함
+            shop.save()
+            #form2.save()
             #return redirect('/shop/{}/'.format(shop.id))
             return redirect(shop)  # get_absolute_url 적용됨
     else:
         form = form_cls(prefix="shop")
-        form2 = form_cls2(prefix="item")
+        #form2 = form_cls2(prefix="item")
 
     return render(request, 'shop/shop_form.html', {
         'form':form,
-        'form2':form2,
+        #'form2':form2,
     })
 
 #shop_new와 동일한 역할 수행함
@@ -72,6 +74,9 @@ def shop_edit(request, pk):
     # except Shop.DoesNotExist:
     #     raise Http404
     shop = get_object_or_404(Shop, pk=pk)  #에러처리 포함된 코드
+
+    if request.user != shop.user: # 최종 작성자와 수정자가 다를 경우
+        return redirect(shop)
 
     form_cls = ShopForm
 
@@ -97,8 +102,11 @@ shop_edit_cbv = UpdateView.as_view(
     model=Shop, form_class=ShopForm) # get_absolute_url 적용됨
 
 @login_required
-def shop_delete(request, pk):
+def shop_delete(request, pk): #(user=request.user)로 하면 미 일치시 404 에러 발생
     shop = get_object_or_404(Shop, pk=pk)
+
+    if request.user != shop.user: # 최종 작성자와 수정자가 다를 경우
+        return redirect(shop)
 
     if request.method == 'POST':
         shop.delete()
